+++ assets/completions/dook.zsh
@@ -30,6 +30,7 @@ _dook () {
 		'--dump=[Dump syntax tree of a file]:file to dump:_files' \
 		'--only-names[Print only names matching the pattern]' \
 		'--offline[Use only already-downloaded parsers]' \
+		'--verbose[Print progress messages]' \
 		{'-h','--help'}'[Print help]' \
 		'*: :{_dook_symbols}'
 }