+++ src/main.rs
@@ -5,9 +5,11 @@
 //     https://dandavison.github.io/delta/grep.html
 //     https://docs.github.com/en/repositories/working-with-files/using-files/navigating-code-on-github#precise-and-search-based-navigation
 
+use enum_derive_2018::EnumFromInner;
 use etcetera::AppStrategy;
 
 mod config;
+mod dep_resolution;
 mod dumptree;
 mod language_name;
 mod loader;
@@ -101,14 +103,22 @@ struct Cli {
     /// Print only names matching the pattern, probably for shell completions.
     #[arg(long)]
     only_names: bool,
+
+    /// Print unstructured messages about progress, for diagnostics.
+    #[arg(short, long, action = clap::ArgAction::Count)]
+    verbose: u8,
 }
 
-#[derive(Debug)]
-enum DookError {
-    IoError(std::io::Error),
-    FileParse(searches::FileParseError),
-    LoaderError(loader::LoaderError),
-    HomeDirError(etcetera::HomeDirError),
+macro_attr_2018::macro_attr! {
+    #[derive(Debug, EnumFromInner!)]
+    enum DookError {
+        IoError(std::io::Error),
+        FileParse(searches::FileParseError),
+        LoaderError(loader::LoaderError),
+        HomeDirError(etcetera::HomeDirError),
+        RipGrepError(RipGrepError),
+        PagerWriteError(PagerWriteError),
+    }
 }
 
 impl std::fmt::Display for DookError {
@@ -118,42 +128,24 @@ impl std::fmt::Display for DookError {
             DookError::IoError(e) => write!(f, "{}", e),
             DookError::LoaderError(e) => write!(f, "{}", e),
             DookError::HomeDirError(e) => write!(f, "{}", e),
+            DookError::RipGrepError(e) => write!(f, "{}", e),
+            DookError::PagerWriteError(e) => write!(f, "{}", e),
         }
-    }
-}
-impl From<std::io::Error> for DookError {
-    fn from(value: std::io::Error) -> Self {
-        Self::IoError(value)
-    }
-}
-
-impl From<searches::FileParseError> for DookError {
-    fn from(value: searches::FileParseError) -> Self {
-        Self::FileParse(value)
-    }
-}
-
-impl From<&searches::FileParseError> for DookError {
-    fn from(value: &searches::FileParseError) -> Self {
-        Self::FileParse(value.clone())
-    }
-}
-
-impl From<loader::LoaderError> for DookError {
-    fn from(value: loader::LoaderError) -> Self {
-        Self::LoaderError(value)
-    }
-}
-
-impl From<etcetera::HomeDirError> for DookError {
-    fn from(value: etcetera::HomeDirError) -> Self {
-        Self::HomeDirError(value)
     }
 }
 
 fn main() -> Result<std::process::ExitCode, DookError> {
+    match main_inner() {
+        // if stdout is gone, let's just leave quietly
+        Err(DookError::PagerWriteError(PagerWriteError::BrokenPipe(_))) => {
+            Ok(std::process::ExitCode::SUCCESS)
+        }
+        result => result,
+    }
+}
+
+fn main_inner() -> Result<std::process::ExitCode, DookError> {
     use clap::Parser;
-    use std::io::Write;
 
     env_logger::init();
 
@@ -193,6 +185,8 @@ fn main() -> Result<std::process::ExitCode, DookError> {
         );
         return Ok(std::process::ExitCode::SUCCESS);
     }
+
+    // get pattern
     let mut current_pattern = cli
         .pattern
         .as_ref()
@@ -203,14 +197,39 @@ fn main() -> Result<std::process::ExitCode, DookError> {
             )
         })?
         .to_owned();
+    // store previous patterns to break --recurse cycles
     let mut local_patterns: std::vec::Vec<regex::Regex> = vec![];
 
-    // store the results here
-    let mut print_ranges: Vec<(Option<std::path::PathBuf>, range_union::RangeUnion)> = Vec::new();
-    let mut print_names: Vec<String> = Vec::new();
+    // set up output
+    let enable_paging = if cli.paging != EnablementLevel::Auto {
+        cli.paging == EnablementLevel::Always
+    } else {
+        cli.plain < 2 && console::Term::stdout().is_term()
+    };
+    let mut pager = paging::MaybePager::new(enable_paging, cli.wrap == WrapMode::Never);
+    let bat_size = console::Term::stdout().size_checked();
+
+    // deduplicate names found under --only-names
+    let mut print_names: std::collections::HashSet<String> = std::collections::HashSet::new();
+
     // parse stdin only once, and upfront, if asked to read it
+    let parse_start = std::time::Instant::now();
     let stdin = parse_stdin(&cli, &mut language_loader, &merged_config);
     let use_stdin = stdin.is_some();
+    if use_stdin && cli.verbose >= 1 {
+        if let Some(stdin) = stdin.as_ref() {
+            write_output_line(
+                &mut pager,
+                format!(
+                    "V: parsed stdin as {:?} in {:?}",
+                    stdin.parsed.as_ref().unwrap().language_name,
+                    parse_start.elapsed(),
+                )
+                .as_bytes(),
+            )?;
+        }
+    }
+
     for is_first_loop in std::iter::once(true).chain(std::iter::repeat(false)) {
         // track recursion
         let mut recurse_defs: std::vec::Vec<String> = vec![];
@@ -223,29 +242,71 @@ fn main() -> Result<std::process::ExitCode, DookError> {
             .last()
             .expect("last() should exist on a vec we just pushed to");
         // pass 0: find candidate files with ripgrep
-        let filenames: Box<dyn Iterator<Item = Option<std::path::PathBuf>>> =
+        if cli.verbose >= 1 {
+            write_output_line(
+                &mut pager,
+                format!("V: invoking ripgrep with {:?}", local_pattern).as_bytes(),
+            )?;
+        }
+        let mut filenames: std::collections::VecDeque<Option<std::path::PathBuf>> =
             if use_stdin && is_first_loop {
-                Box::new(std::iter::once(None))
+                std::collections::VecDeque::from([None])
             } else {
-                let ripgrep_results = ripgrep(&current_pattern)?.into_iter().map(Some);
+                let ripgrep_results = ripgrep(&current_pattern).filter_map(|f| match f {
+                    Ok(p) => Some(Some(p)),
+                    Err(e) => {
+                        eprintln!("{}", e);
+                        None
+                    }
+                });
                 if use_stdin {
-                    Box::new(std::iter::once(None).chain(ripgrep_results))
+                    std::iter::once(None).chain(ripgrep_results).collect()
                 } else {
-                    Box::new(ripgrep_results)
+                    ripgrep_results.collect()
                 }
             };
-        for path in filenames {
-            // infer syntax, then search with tree_sitter
+        if cli.verbose >= 1 {
+            let file_count = if use_stdin {
+                filenames.len().saturating_sub(1)
+            } else {
+                filenames.len()
+            };
+            write_output_line(
+                &mut pager,
+                format!("V: ripgrep found {} files", file_count).as_bytes(),
+            )?;
+        }
+        // track import origins seen so far
+        let mut import_origins: std::collections::HashSet<String> =
+            std::collections::HashSet::new();
+        loop {
+            let path = match filenames.pop_front() {
+                None => break,
+                Some(path) => path,
+            };
+            if cli.verbose >= 1 {
+                match &path {
+                    None => write_output_line(&mut pager, "V: parsing stdin".as_bytes())?,
+                    Some(path) => {
+                        write_output_line(&mut pager, format!("V: parsing {:?}", path).as_bytes())?
+                    }
+                }
+            }
+            // infer syntax
+            let parse_start = std::time::Instant::now();
             let tmp_parsed_file: searches::ParsedFile; // storage for file_info if created on the fly
-            let file_info = match path {
-                None => stdin.as_ref().expect("oops we weren't given --stdin but somehow we queued stdin to search anyway").parsed.as_ref()?,
+            let file_info = match &path {
+                None => stdin.as_ref()
+                    .expect("oops we weren't given --stdin but somehow we queued stdin to search anyway")
+                    .parsed.as_ref().map_err(|e| e.clone())?,
                 Some(path) => match searches::ParsedFile::from_filename(
-                    &path,
+                    path,
                     &mut language_loader,
                     &merged_config,
                 ) {
                     Err(e) => {
-                        eprintln!("Skipping {:?}: {}", &path, e);
+                        let error_message = format!("Skipping {:?}: {}", &path, e);
+                        write_output_line(&mut pager, error_message.as_bytes())?;
                         continue;
                     }
                     Ok(f) => {
@@ -254,22 +315,39 @@ fn main() -> Result<std::process::ExitCode, DookError> {
                     }
                 },
             };
+            if cli.verbose >= 1 {
+                write_output_line(
+                    &mut pager,
+                    format!(
+                        "V: parsed as {:?} in {:?}",
+                        file_info.language_name,
+                        parse_start.elapsed()
+                    )
+                    .as_bytes(),
+                )?;
+            }
+            // get corresponding search patterns
             let language_info = match query_compiler
                 .get_language_info(file_info.language_name, &mut language_loader)
             {
                 Ok(language_info) => language_info,
                 Err(e) => {
-                    eprintln!("{}", e);
+                    write_output_line(&mut pager, e.to_string().as_bytes())?;
                     continue;
                 }
             };
+            // search with tree_sitter
             if cli.only_names {
-                print_names.extend_from_slice(&searches::find_names(
+                for name in searches::find_names(
                     file_info.source_code.as_slice(),
                     &file_info.tree,
                     &language_info,
                     local_pattern,
-                ));
+                ) {
+                    if print_names.insert(name.clone()) {
+                        write_output_line(&mut pager, name.as_bytes())?;
+                    }
+                }
             } else {
                 let search_result = searches::find_definition(
                     file_info.source_code.as_slice(),
@@ -279,14 +357,72 @@ fn main() -> Result<std::process::ExitCode, DookError> {
                     true,
                 );
                 if !search_result.ranges.is_empty() {
-                    print_ranges.push((file_info.path.to_owned(), search_result.ranges)); // TODO extend prev if new_ranges comes after in the same file
+                    // It could be nice to do a single bat invocation in the
+                    // rare case that consecutive recursions hit the same file,
+                    // but printing results as they come gets in the way.
+                    match write_ranges(
+                        &mut pager,
+                        path.as_ref(),
+                        &search_result.ranges,
+                        &cli,
+                        use_color,
+                        bat_size,
+                        stdin.as_ref(),
+                    ) {
+                        // if stdout is gone, just leave quietly
+                        Err(PagerWriteError::BrokenPipe(_)) => {
+                            return Ok(std::process::ExitCode::SUCCESS)
+                        }
+                        _ if pager.is_abandoned() => return Ok(std::process::ExitCode::SUCCESS),
+                        // for other errors, print and continue
+                        Err(e) => {
+                            let error_message = format!("Error reading {:?}: {}", path, e);
+                            pager.write_line(error_message.as_bytes())?;
+                        }
+                        Ok(_) => (),
+                    }
                     recurse_defs.extend(search_result.recurse_names.into_iter().filter(|name| {
                         local_patterns.iter().all(|pattern| !pattern.is_match(name))
                     }));
                 }
+
+                // follow probable imports if we know about them
+                for import_pattern in search_result.import_origins {
+                    if import_origins.insert(import_pattern.clone()) {
+                        if cli.verbose >= 1 {
+                            write_output_line(
+                                &mut pager,
+                                format!(
+                                    "V: sorting files matching {:?} to the front",
+                                    import_pattern
+                                )
+                                .as_bytes(),
+                            )?;
+                        }
+
+                        filenames
+                            .make_contiguous()
+                            .sort_by_cached_key(|path| match path {
+                                None => 0,
+                                Some(path) => dep_resolution::dissimilarity(
+                                    file_info.language_name,
+                                    &import_pattern,
+                                    path,
+                                ),
+                            });
+                    }
+                }
             }
         }
+
+        // recursion
         recurse_defs.dedup();
+        if cli.verbose >= 1 {
+            write_output_line(
+                &mut pager,
+                format!("V: recursion candidates: {:?}", recurse_defs).as_bytes(),
+            )?;
+        }
         if cli.recurse && !cli.only_names && recurse_defs.len() == 1 {
             current_pattern = regex::Regex::new(&regex::escape(&recurse_defs[0])).unwrap();
         } else {
@@ -294,98 +430,6 @@ fn main() -> Result<std::process::ExitCode, DookError> {
         }
     }
 
-    // early exit for --only-names
-    if cli.only_names {
-        print_names.sort();
-        print_names.dedup();
-        let stdout = console::Term::stdout();
-        for name in print_names {
-            stdout.write_line(&name)?;
-        }
-        return Ok(std::process::ExitCode::SUCCESS);
-    }
-
-    // set up paging if requested
-    let enable_paging = if cli.paging != EnablementLevel::Auto {
-        cli.paging == EnablementLevel::Always
-    } else {
-        cli.plain < 2 && console::Term::stdout().is_term()
-    };
-    let mut pager = paging::MaybePager::new(enable_paging, cli.wrap == WrapMode::Never);
-    let bat_size = console::Term::stdout().size_checked();
-    for (path, ranges) in print_ranges.iter() {
-        let mut cmd = std::process::Command::new("bat");
-        let cmd = cmd
-            .arg("--paging=never")
-            .arg(format!("--wrap={:?}", cli.wrap).to_lowercase())
-            .arg(format!("--color={:?}", use_color).to_lowercase());
-        let cmd = match bat_size {
-            Some((_rows, cols)) => cmd.arg(format!("--terminal-width={}", cols)),
-            None => cmd,
-        };
-        let cmd = match cli.plain {
-            0 => cmd,
-            _ => cmd.arg("--plain"),
-        };
-        let cmd = cmd
-            .args(
-                ranges
-                    .iter_filling_gaps(1) // snip indicator - 8< - takes 1 line anyway
-                    .map(|x| format!("--line-range={}:{}", x.start + 1, x.end)), // bat end is inclusive
-            )
-            .stderr(std::process::Stdio::inherit())
-            .stdout(std::process::Stdio::piped());
-        let child = match path {
-            Some(path) => cmd.arg(path).spawn(),
-            None => {
-                let stdin = stdin.as_ref().expect("oops we weren't given --stdin but somehow we claim we have results from stdin anyway");
-                let mut child = cmd
-                    .arg(format!(
-                        "-l{}",
-                        stdin.parsed.as_ref().unwrap().language_name_str
-                    ))
-                    .stdin(std::process::Stdio::piped())
-                    .spawn();
-                if let Ok(child) = &mut child {
-                    let mut child_stdin = child.stdin.take().unwrap();
-                    let stdin_clone = stdin.bytes.clone();
-                    std::thread::spawn(move || {
-                        let result = child_stdin.write_all(&stdin_clone);
-                        if paging::is_broken_pipe(&result) {
-                        } else {
-                            result.unwrap()
-                        }
-                    });
-                }
-                child
-            }
-        };
-        if let Err(e) = child.and_then(|mut child| {
-            // std::io::copy uses more efficient syscalls than explicitly reading and writing
-            let mut child_stdout = child.stdout.take().unwrap();
-            let copy_result = std::io::copy(&mut child_stdout, &mut pager);
-            let wait_result = child.wait().and_then(|exit_status| {
-                if exit_status.success() {
-                    Ok(())
-                } else {
-                    Err(std::io::Error::other(format!(
-                        "reader exited {:?}",
-                        exit_status
-                    )))
-                }
-            });
-            copy_result.map(|_| wait_result)
-        }) {
-            if pager.is_abandoned() {
-                // stdout is gone so let's just leave quietly
-                return Ok(std::process::ExitCode::SUCCESS);
-            } else {
-                let error_message =
-                    std::vec::Vec::from(format!("Error reading {:?}: {}\n", path, e));
-                pager.write_all(&error_message)?;
-            }
-        }
-    }
     // wait for pager
     match pager.wait() {
         Ok(0) => (),
@@ -395,6 +439,124 @@ fn main() -> Result<std::process::ExitCode, DookError> {
 
     // yeah yeah whatever
     Ok(std::process::ExitCode::SUCCESS)
+}
+
+#[derive(Debug)]
+enum PagerWriteError {
+    IoError(std::io::Error),
+    BrokenPipe(()),
+    ReaderDied(std::process::ExitStatus),
+}
+
+impl std::fmt::Display for PagerWriteError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            PagerWriteError::IoError(e) => write!(f, "{}", e),
+            PagerWriteError::BrokenPipe(_) => write!(f, "broken pipe (someone closed our output)"),
+            PagerWriteError::ReaderDied(status) => {
+                write!(f, "formatting excerpt exited {}", status)
+            }
+        }
+    }
+}
+
+impl From<std::io::Error> for PagerWriteError {
+    fn from(value: std::io::Error) -> Self {
+        match value.kind() {
+            std::io::ErrorKind::BrokenPipe => Self::BrokenPipe(()),
+            _ => Self::IoError(value),
+        }
+    }
+}
+
+fn write_output_line(pager: &mut paging::MaybePager, line: &[u8]) -> Result<(), PagerWriteError> {
+    match pager.write_line(line) {
+        Ok(x) => Ok(x),
+        Err(e) => {
+            if pager.is_abandoned() {
+                Err(PagerWriteError::BrokenPipe(()))
+            } else {
+                Err(e)?
+            }
+        }
+    }
+}
+
+// this function signature is a disaster
+fn write_ranges(
+    pager: &mut paging::MaybePager,
+    path: Option<&std::path::PathBuf>,
+    ranges: &range_union::RangeUnion,
+    cli: &Cli,
+    use_color: EnablementLevel,
+    bat_size: Option<(u16, u16)>,
+    stdin: Option<&Stdin>,
+) -> Result<(), PagerWriteError> {
+    use std::io::Write;
+
+    let mut cmd = std::process::Command::new("bat");
+    let cmd = cmd
+        .arg("--paging=never")
+        .arg(format!("--wrap={:?}", cli.wrap).to_lowercase())
+        .arg(format!("--color={:?}", use_color).to_lowercase());
+    let cmd = match bat_size {
+        Some((_rows, cols)) => cmd.arg(format!("--terminal-width={}", cols)),
+        None => cmd,
+    };
+    let cmd = match cli.plain {
+        0 => cmd,
+        _ => cmd.arg("--plain"),
+    };
+    let cmd = cmd
+        .args(
+            ranges
+                .iter_filling_gaps(1) // snip indicator - 8< - takes 1 line anyway
+                .map(|x| format!("--line-range={}:{}", x.start + 1, x.end)), // bat end is inclusive
+        )
+        .stderr(std::process::Stdio::inherit())
+        .stdout(std::process::Stdio::piped());
+    let mut child = match path {
+        Some(path) => cmd.arg(path).spawn(),
+        None => {
+            let stdin = stdin.expect("oops we weren't given --stdin but somehow we claim we have results from stdin anyway");
+            let mut child = cmd
+                .arg(format!(
+                    "-l{}",
+                    stdin.parsed.as_ref().unwrap().language_name_str
+                ))
+                .stdin(std::process::Stdio::piped())
+                .spawn();
+            if let Ok(child) = &mut child {
+                let mut child_stdin = child
+                    .stdin
+                    .take()
+                    .expect("BUG: should have launched bat with stdin=piped()");
+                let stdin_clone = stdin.bytes.clone();
+                std::thread::spawn(move || {
+                    let result = child_stdin.write_all(&stdin_clone);
+                    if paging::is_broken_pipe(&result) {
+                    } else {
+                        result.unwrap()
+                    }
+                });
+            }
+            child
+        }
+    }?;
+    // std::io::copy uses more efficient syscalls than explicitly reading and writing
+    let mut child_stdout = child
+        .stdout
+        .take()
+        .expect("BUG: should have launched bat with stdout=piped()");
+    let copy_result = std::io::copy(&mut child_stdout, pager);
+    let wait_result = child.wait();
+    copy_result?;
+    let exit_status = wait_result?;
+    if exit_status.success() {
+        Ok(())
+    } else {
+        Err(PagerWriteError::ReaderDied(exit_status))
+    }
 }
 
 struct Stdin {
@@ -432,51 +594,59 @@ fn parse_stdin(
     Some(Stdin { bytes, parsed })
 }
 
-fn ripgrep(pattern: &regex::Regex) -> std::io::Result<Vec<std::path::PathBuf>> {
+#[derive(Debug)]
+enum RipGrepError {
+    NotLaunched(std::io::Error),
+    ReadFailed(std::io::Error),
+    FileNameUnparseable(Vec<u8>),
+}
+
+#[rustfmt::skip] // keep compact
+impl std::fmt::Display for RipGrepError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            RipGrepError::NotLaunched(e)
+                => write!(f, "failed to run ripgrep: {}", e),
+            RipGrepError::ReadFailed(e)
+                => write!(f, "failed to read ripgrep output: {}", e),
+            RipGrepError::FileNameUnparseable(filename)
+                => write!(f, "ripgrep returned unreadable filename: {:?}", filename),
+        }
+    }
+}
+
+fn ripgrep(
+    pattern: &regex::Regex,
+) -> Box<dyn Iterator<Item = Result<std::path::PathBuf, RipGrepError>>> {
     use os_str_bytes::OsStrBytes;
+    use std::io::BufRead;
 
     // first-pass search with ripgrep
     let mut rg = std::process::Command::new("rg");
-    let rg_output = rg
-        .arg("-l")
-        .arg("-0")
+    let mut child = match rg
+        .args(["-l", "--sort=path", "-0"])
         .arg(pattern.as_str())
         .arg("./")
         .stderr(std::process::Stdio::inherit())
-        .output()?;
-    if !rg_output.status.success() {
-        if let Some(e) = rg_output.status.code() {
-            return Err(std::io::Error::new(
-                // TODO adopt this exit code as our own
-                std::io::ErrorKind::Other,
-                format!("ripgrep exited {:?}", e),
-            ));
-        }
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::Other,
-            format!("{}", rg_output.status),
-        ));
-    }
-    // TODO is this even actually the right way to convert stdout to OsStr?
-    let filenames: std::io::Result<Vec<std::path::PathBuf>> = rg_output
-        .stdout
-        .split(|x| *x == 0)
-        .filter_map(|x| match std::ffi::OsStr::from_io_bytes(x) {
-            None => Some(Err(std::io::Error::new(
-                std::io::ErrorKind::InvalidInput,
-                format!("{:?}", std::vec::Vec::from(x)),
-            ))),
+        .stdout(std::process::Stdio::piped())
+        .spawn()
+    {
+        Ok(c) => c,
+        Err(e) => return Box::new(std::iter::once(Err(RipGrepError::NotLaunched(e)))),
+    };
+    let child_stdout = child.stdout.take().unwrap();
+    let rg_lines = std::io::BufReader::new(child_stdout).split(0);
+    Box::new(rg_lines.filter_map(|x| match x {
+        Err(e) => Some(Err(RipGrepError::ReadFailed(e))),
+        Ok(x) => match std::ffi::OsStr::from_io_bytes(&x) {
+            None => Some(Err(RipGrepError::FileNameUnparseable(x))),
             Some(y) => {
                 if y.is_empty() {
                     None
                 } else {
-                    Some(Ok(y.to_owned().into()))
+                    Some(Ok(std::path::PathBuf::from(y)))
                 }
             }
-        })
-        .collect();
-    filenames.map(|mut f| {
-        f.sort_unstable();
-        f
-    })
+        },
+    }))
 }